webserver/requirements.txt
asgiref==3.3.1
Django==3.1.6
djangorestframework==3.12.2
Pillow==8.1.0
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
